Rename package from source-auditor to SPDX
Preparing to move to the SPDX organization
