+++ CONTRIBUTING.md
@@ -4,13 +4,13 @@ Thank you for your interest in `spdx-to-osv`. The project is open-source softwar
 
 Issues
 ------
-`spdx-to-osv` has a [project page on GitHub](https://github.com/goneall/spdx-to-osv) where you can [create an issue](https://github.com/goneall/spdx-to-osv/issues/new/choose) to report a bug, make a suggestion, or propose a substantial change or improvement that you might like to make.
+`spdx-to-osv` has a [project page on GitHub](https://github.com/spdx/spdx-to-osv) where you can [create an issue](https://github.com/spdx/spdx-to-osv/issues/new/choose) to report a bug, make a suggestion, or propose a substantial change or improvement that you might like to make.
 
 If you would like to work on a fix for any issue, please assign the issue to yourself prior to creating a Pull Request.
 
 Pull Requests
 -------
-The source code for `spdx-to-osv` is hosted on [github.com/goneall/spdx-to-osv](https://github.com/goneall/spdx-to-osv). Please review [open pull requests](https://github.com/goneall/spdx-to-osv/pulls) and [active branches](https://github.com/goneall/spdx-to-osv/branches) before committing time to a substantial revision. Work along similar lines may already be in progress.
+The source code for `spdx-to-osv` is hosted on [github.com/spdx/spdx-to-osv](https://github.com/spdx/spdx-to-osv). Please review [open pull requests](https://github.com/spdx/spdx-to-osv/pulls) and [active branches](https://github.com/spdx/spdx-to-osv/branches) before committing time to a substantial revision. Work along similar lines may already be in progress.
 
 To submit a pull request via GitHub, fork the repository, create a topic branch from `main` for your work, and send a pull request when ready. If you would prefer to send a patch or grant access to pull from your own Git repository, please contact the project's contributors by e-mail.
