+++ CREDITS.md
@@ -2,3 +2,7 @@
 - This utility uses the [SPDX Tools Java](https://github.com/spdx/tools-java) maintained by the SPDX community licensed under the Apache-2.0 license for the SPDX parsing
 
 - This utility uses the [CPE-parser](https://github.com/stevespringett/CPE-Parser) from Steve Springett licensed under the Apache-2.0 license for parsing CPE strings
+
+- The utility was originally created by Gary O'Neall (Source Auditor Inc.)
+
+- The utility includes contributions from Brandon Lum (Google Inc.)
