--- src/main/java/com/sourceauditor/spdx_to_osv/SwhException.java
@@ -1,58 +1,0 @@
-/**
- * SPDX-License-Identifier: Apache-2.0
- * Copyright (c) 2021 Source Auditor Inc.
- */
-package com.sourceauditor.spdx_to_osv;
-
-/**
- * Exceptions related to the processing of Software Heritage
- * @author Gary O'Neall
- *
- */
-public class SwhException extends Exception {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-	/**
-	 * 
-	 */
-	public SwhException() {
-	}
-
-	/**
-	 * @param arg0
-	 */
-	public SwhException(String arg0) {
-		super(arg0);
-	}
-
-	/**
-	 * @param arg0
-	 */
-	public SwhException(Throwable arg0) {
-		super(arg0);
-	}
-
-	/**
-	 * @param arg0
-	 * @param arg1
-	 */
-	public SwhException(String arg0, Throwable arg1) {
-		super(arg0, arg1);
-	}
-
-	/**
-	 * @param arg0
-	 * @param arg1
-	 * @param arg2
-	 * @param arg3
-	 */
-	public SwhException(String arg0, Throwable arg1, boolean arg2,
-			boolean arg3) {
-		super(arg0, arg1, arg2, arg3);
-	}
-
-}
