+++ src/main/java/org/spdx/spdx_to_osv/osvmodel/OsvVulnerabilityResponse.java
@@ -1,0 +1,23 @@
+/**
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) 2021 Source Auditor Inc.
+ */
+package org.spdx.spdx_to_osv.osvmodel;
+
+import java.util.List;
+
+/**
+ * Object for a response from the OSV-QueryAffected API based on https://osv.dev/docs/#operation/OSV_QueryAffected
+ */
+public class OsvVulnerabilityResponse {
+
+    List<OsvVulnerability> vulns = null;
+    
+    public OsvVulnerabilityResponse() {
+        // required empty constructor
+    }
+    
+    public List<OsvVulnerability> getVulns() {
+        return this.vulns;
+    }
+}
